python/requirements.txt
# Core dependencies
web3>=6.0.0
solana>=0.30.0
solders>=0.18.0
anchorpy>=0.18.0
eth-account>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Ethereum/Polygon/BNB Chain
eth-abi>=4.0.0
eth-utils>=2.3.0
eth-typing>=3.5.0

# Solana
base58>=2.1.1
solana-py>=0.30.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
typing-extensions>=4.8.0

# API clients
httpx>=0.25.0
